refactor messages on home and job page and in general
